+++ src/LAS.h
@@ -28,7 +28,7 @@ class LAS
     void filter_in_polygon(std::string wkt);
     void filter_local_maxima(NumericVector ws, double min_height, bool circular);
     void filter_local_maxima(NumericVector ws);
-    void filter_with_grid(S4 layout);
+    void filter_with_grid(S4 layout, bool max);
     void filter_shape(int method, NumericVector th, int k);
     void filter_progressive_morphology(NumericVector ws, NumericVector th);
     void filter_isolated_voxel(double ws, unsigned int isolated);